requirements.txt
langchain>=0.3.23
langchain-community>=0.3.21
langchain-core>=0.3.51
langchain-ollama>=0.3.1
crewai>=0.114.0
langchain-openai>=0.3.12
prophet>=1.1.6
torch>=2.2.0
pandas>=2.1.1
numpy>=1.26.0
scikit-learn>=1.3.2
transformers>=4.34.0
chromadb>=0.4.18
fastapi>=0.103.2
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
boto3>=1.28.62
psycopg2-binary>=2.9.9
matplotlib>=3.8.0
plotly>=5.17.0
pytest>=7.4.2
confluent-kafka>=2.2.0
prometheus-client>=0.17.1 